requirements.txt
python-telegram-bot==21.6
python-dotenv==1.0.1
requests==2.32.3
easyocr==1.7.2
pillow==10.4.0
scikit-learn==1.5.1
